requirements.txt
pandas==2.2.2
python-chess==1.999
scikit-learn==1.4.2
numpy==1.26.0
matplotlib==3.8.4
